requirements.txt
pycryptodome>=3.20.0
psutil>=5.9.0
matplotlib>=3.8.0
tabulate>=0.9.0
pandas>=2.0.0
cryptography>=42.0.0
